refactor claim
