+++ coin-vester-front-2/src/const/sc.ts
@@ -1,1 +1,2 @@
-export const scAddr = 'AS14oUcMujECzburUs8iMZZGsJQQQ61MxPVUKyV8WUNrw3mHmSqF';
+export const SC_ADDRESS =
+  'AS14oUcMujECzburUs8iMZZGsJQQQ61MxPVUKyV8WUNrw3mHmSqF';
